refactor(cru-users): fix theo comment cua trainer va sua them logic cua updateUser
